move list and edit routes out of active route scope
