+++ extension/src/panel/pages/ConnectionStack/index.tsx
