+++ extension/src/panel/pages/index.tsx
@@ -1,5 +1,6 @@
 import { type RouteObject } from 'react-router'
 import { ActiveRoute } from './$activeRouteId'
 import { NoRoutes } from './_index'
+import { routes } from './routes'
 
-export const pages: RouteObject[] = [NoRoutes, ActiveRoute]
+export const pages: RouteObject[] = [NoRoutes, ActiveRoute, routes]
