+++ extension/src/panel/pages/legacyConnectionMigrations.test.ts
