+++ extension/src/panel/pages/legacyConnectionMigrations.ts
