+++ extension/src/panel/pages/routes/edit.$routeId/AvatarInput.tsx
