+++ extension/src/panel/pages/routes/edit.$routeId/RemoveButton.tsx
