+++ extension/src/panel/pages/routes/edit.$routeId/ZodiacMod.tsx
