+++ extension/src/panel/pages/routes/edit.$routeId/index.tsx
