+++ extension/src/panel/pages/routes/edit.$routeId/wallet/WalletDisconnected.tsx
