+++ extension/src/panel/pages/routes/edit.$routeId/wallet/index.ts
