+++ extension/src/panel/pages/routes/edit.$routeId/wallet/injectedWallet/InjectedWallet.tsx
