+++ extension/src/panel/pages/routes/edit.$routeId/wallet/injectedWallet/InjectedWalletConnect.tsx
