+++ extension/src/panel/pages/routes/edit.$routeId/wallet/injectedWallet/index.ts
