+++ extension/src/panel/pages/routes/edit.$routeId/wallet/providerLogo/ProviderLogo.tsx
