+++ extension/src/panel/pages/routes/edit.$routeId/wallet/providerLogo/index.ts
