+++ extension/src/panel/pages/routes/edit.$routeId/wallet/providerLogo/wallet-connect-logo.png
