+++ extension/src/panel/pages/routes/edit.$routeId/wallet/walletConnect/WalletConnect.tsx
