+++ extension/src/panel/pages/routes/edit.$routeId/wallet/walletConnect/WalletConnectConnect.tsx
