+++ extension/src/panel/pages/routes/edit.$routeId/wallet/walletConnect/index.ts
