+++ extension/src/panel/pages/useConfirmClearTransaction.tsx
